archive/old-document-extraction/requirements.txt
flask
python-docx
PyMuPDF
pytesseract
Pillow
spacy
googletrans==4.0.0rc1
pyyaml